Adds a GitHub workflow to deploy frozen site to gh-pages
Inspired by the viral wget workflow here:

https://github.com/pubstandards/pubstandards-london/blob/89989bd3e3e2391425bc8e76055e90892635ec5a/.github/workflows/build.yml

Already using freeze to deploy this site, so much simpler. Haven't
bothered with gh-pages and instead have been pushing to a tiny Digital
Ocean droplet so this may be a nice way to save $5 a month -- coffee
money! (Wait what are coffee shops again? It's a pandemic and I forgot)